+++ src/app/server/services/collectionsService.js
@@ -1,6 +1,11 @@
+import createHttpError from 'http-errors';
+import { saveFileToUploadDir } from '../lib';
+import { saveFileToCloudinary } from '../lib/saveFileToCloudinary';
 import CollectionModel from '../models/CollectionsModel';
+import { env } from '../utils';
+import mongoose from 'mongoose';
 
-export const getAllCollections = async () => {
+export const getAllCollections = async (req, res) => {
   try {
     const collections = await CollectionModel.find().lean();
 
@@ -15,49 +20,105 @@ export const getAllCollections = async () => {
       },
       { activeCollections: [], closedCollections: [] }
     );
-
-    return sortedCollections;
+    res.status(200).json({ status: 200, data: sortedCollections });
   } catch (error) {
-    throw new Error('Failed to retrieve collections');
+    res.status(500).json({ message: error.message });
   }
 };
 
-export const getCollectionById = async (id) => {
-  const collection = await CollectionModel.findOne({
-    _id: id,
-  });
-
-  return collection;
-};
-
-export const createCollection = async (payload) => {
-  const newCollection = new CollectionModel(payload);
-  await newCollection.save();
-  return newCollection;
-};
-
-export const updateCollectionService = async (id, payload, options = {}) => {
-  if (!id) {
-    return null;
-  }
-
+export const getCollectionById = async (req, res) => {
   try {
-    const updatedCollection = await CollectionModel.findByIdAndUpdate(
+    const { id } = req.query;
+
+    const collection = await CollectionModel.findOne({
+      _id: id,
+    });
+    if (!collection) {
+      return res.status(404).json({ error: 'Collection not found' });
+    }
+    res.status(200).json({ status: 200, data: collection });
+  } catch (error) {
+    res.status(500).json({ error: 'Failed to fetch collection' });
+  }
+};
+
+export const createCollection = async (req, res) => {
+  try {
+    const image = req.files;
+    let photoUrls = [];
+
+    if (image && image.length) {
+      for (const img of image) {
+        let photoUrl;
+        if (env('ENABLE_CLOUDINARY') === 'true') {
+          photoUrl = await saveFileToCloudinary(img);
+        } else {
+          photoUrl = await saveFileToUploadDir(img);
+        }
+        photoUrls.push(photoUrl);
+      }
+    }
+
+    const newCollection = new CollectionModel({
+      ...req.body,
+      image: photoUrls,
+    });
+    await newCollection.save();
+    res.status(201).json(newCollection);
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
+};
+
+export const updateCollectionService = async (req, res) => {
+  try {
+    const { id } = req.query;
+    const image = req.files;
+
+    if (!id) {
+      return null;
+    }
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: 'Invalid ID format' });
+    }
+    const existingCollection = await CollectionModel.findById(id);
+
+    if (!existingCollection) {
+      throw createHttpError(404, 'Collection not found');
+    }
+
+    let photoUrls = existingCollection.image || [];
+
+    if (image && image.length) {
+      for (const img of image) {
+        let photoUrl;
+        if (env('ENABLE_CLOUDINARY') === 'true') {
+          photoUrl = await saveFileToCloudinary(img);
+        } else {
+          photoUrl = await saveFileToUploadDir(img);
+        }
+        photoUrls.push(photoUrl);
+      }
+    }
+
+    const result = await CollectionModel.findByIdAndUpdate(
       { _id: id },
-      payload,
+      { ...req.body, image: photoUrls },
       {
         new: true,
         includeResultMetadata: true,
-        ...options,
       }
     );
 
-    if (!updatedCollection) {
-      return null; 
+    if (!result) {
+      throw createHttpError(404, 'Collection not found');
     }
 
-    return updatedCollection.value; 
+    res.status(200).json({
+      data: result.value,
+    });
   } catch (error) {
-    return null; 
+    res.status(500).json({ message: error.message });
   }
 };
