+++ src/app/server/services/reportingsService.js
@@ -1,51 +1,118 @@
+import mongoose from 'mongoose';
+import { saveFileToUploadDir } from '../lib';
+import { saveFileToCloudinary } from '../lib/saveFileToCloudinary';
 import ReportingsModel from '../models/ReportingModels';
+import { env } from '../utils';
+import createHttpError from 'http-errors';
 
-export const getAllReportings = async () => {
+export const getAllReportings = async (req, res) => {
   try {
     const reportings = await ReportingsModel.find().lean();
 
-    return reportings;
+    res.status(200).json({ status: 200, data: reportings });
   } catch (error) {
-    throw new Error('Failed to retrieve reportings');
+    res.status(500).json({ message: error.message });
   }
 };
 
-export const getReportingById = async (id) => {
-  const collection = await ReportingsModel.findOne({
-    _id: id,
-  });
-
-  return collection;
-};
-
-export const createReporting = async (payload) => {
-  const newReport = new ReportingsModel(payload);
-  await newReport.save();
-  return newReport;
-};
-
-export const updateReportingService = async (id, payload, options = {}) => {
-  if (!id) {
-    return null;
-  }
-
+export const getReportingById = async (req, res) => {
   try {
-    const updatedReport = await ReportingsModel.findByIdAndUpdate(
-      { _id: id },
-      payload,
-      {
-        new: true,
-        includeResultMetadata: true,
-        ...options,
-      }
-    );
+    const { id } = req.query;
 
-    if (!updatedReport) {
+    const report = await ReportingsModel.findOne({
+      _id: id,
+    });
+
+    if (!report) {
+      return res.status(404).json({ error: 'Report not found' });
+    }
+
+    res.status(200).json(report);
+  } catch (error) {
+    res.status(500).json({ error: 'Failed to fetch report' });
+  }
+};
+
+export const createReporting = async (req, res) => {
+  try {
+    const image = req.files;
+    let photoUrls = [];
+
+    if (image && image.length) {
+      for (const img of image) {
+        let photoUrl;
+        if (env('ENABLE_CLOUDINARY') === 'true') {
+          photoUrl = await saveFileToCloudinary(img);
+        } else {
+          photoUrl = await saveFileToUploadDir(img);
+        }
+        photoUrls.push(photoUrl);
+      }
+    }
+
+    const newReport = new ReportingsModel({
+      ...req.body,
+      image: photoUrls,
+    });
+
+    await newReport.save();
+
+    res.status(201).json(newReport);
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
+};
+
+export const updateReportingService = async (req, res) => {
+  try {
+    const { id } = req.query;
+    const image = req.files;
+    if (!id) {
       return null;
     }
 
-    return updatedReport.value;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(400).json({ message: 'Invalid ID format' });
+    }
+    const existingReport = await ReportingsModel.findById(id);
+
+    if (!existingReport) {
+      throw createHttpError(404, 'Report not found');
+    }
+
+    let photoUrls = existingReport.image || [];
+
+    if (image && image.length) {
+      for (const img of image) {
+        let photoUrl;
+        if (env('ENABLE_CLOUDINARY') === 'true') {
+          photoUrl = await saveFileToCloudinary(img);
+        } else {
+          photoUrl = await saveFileToUploadDir(img);
+        }
+        photoUrls.push(photoUrl);
+      }
+    }
+
+    const result = await ReportingsModel.findByIdAndUpdate(
+      id,
+      {
+        ...req.body,
+        image: photoUrls,
+      },
+      { new: true, includeResultMetadata: true }
+    );
+
+    if (!result) {
+      throw createHttpError(404, 'Report not found');
+    }
+
+    res.json({
+      status: 200,
+      message: 'Successfully patched a report!',
+      data: result.value,
+    });
   } catch (error) {
-    return null;
+    res.status(500).json({ message: error.message });
   }
 };
