+++ src/pages/api/all-pages/en/[route].js
@@ -1,5 +1,5 @@
-import { handleGetByPoutPageEN } from '@/app/server/controllers/pagesController';
 import { connectToDatabase } from '@/app/server/lib';
+import { getPageByRouteEN } from '@/app/server/services/pagesServices';
 
 export const config = {
   api: {
@@ -11,7 +11,7 @@ export default async function handler(req, res) {
   await connectToDatabase();
 
   if (req.method === 'GET') {
-    return handleGetByPoutPageEN(req, res);
+    return getPageByRouteEN(req, res);
   }
 
   res.status(405).json({ message: 'Method Not Allowed' });
