+++ src/pages/api/all-pages/uk/[route].js
@@ -1,5 +1,5 @@
-import { handleGetByPoutPageUK } from '@/app/server/controllers/pagesController';
 import { connectToDatabase } from '@/app/server/lib';
+import { getPageByRouteUK } from '@/app/server/services/pagesServices';
 
 export const config = {
   api: {
@@ -11,7 +11,7 @@ export default async function handler(req, res) {
   await connectToDatabase();
 
   if (req.method === 'GET') {
-    return handleGetByPoutPageUK(req, res);
+    return getPageByRouteUK(req, res);
   }
 
   res.status(405).json({ message: 'Method Not Allowed' });
