+++ src/pages/api/collections/[id].js
@@ -1,12 +1,12 @@
 import connectToDatabase from '@/app/server/lib/mongodb.js';
 import { uploadMiddleware } from '../../../app/server/lib/multer.js';
-import {
-  getCollectionByIdController,
-  handleUpdateCollection,
-} from '@/app/server/controllers/collectionsController.js';
 import isValidId from '@/app/server/utils/isValidId.js';
 import validateBody from '@/app/server/utils/validateBody.js';
 import CollectionModel from '@/app/server/models/CollectionsModel.js';
+import {
+  getCollectionById,
+  updateCollectionService,
+} from '@/app/server/services/collectionsService.js';
 export const config = {
   api: {
     bodyParser: false,
@@ -17,14 +17,14 @@ export default async function handler(req, res) {
   await connectToDatabase();
 
   if (req.method === 'GET') {
-    return getCollectionByIdController(req, res);
+    return getCollectionById(req, res);
   }
 
   if (req.method === 'PATCH') {
     isValidId(req, res);
     await uploadMiddleware(req, res);
     validateBody(CollectionModel);
-    return handleUpdateCollection(req, res);
+    return updateCollectionService(req, res);
   }
 
   res.status(405).json({ message: 'Method not allowed' });
